pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdfx-bench"
version = "1.0.0"
description = "No-hallucination PDF extraction benchmark"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "PDFX-Bench Team", email = "team@pdfx-bench.com"}
]
keywords = ["pdf", "extraction", "benchmark", "tables", "text"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Text Processing :: Markup",
]

dependencies = [
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "PyMuPDF>=1.23.0",
    "pdfplumber>=0.10.0",
    "camelot-py[cv]>=0.11.0",
    "tabula-py>=2.8.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
cloud = [
    "boto3>=1.28.0",
    "google-cloud-documentai>=2.20.0",
    "azure-ai-documentintelligence>=1.0.0",
    "pdfservices-sdk>=4.0.0",
]
ocr = [
    "pytesseract>=0.3.10",
    "pdf2image>=3.1.0",
]
export = [
    "openpyxl>=3.1.0",
    "pyarrow>=12.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
all = [
    "pdfx-bench[cloud,ocr,export,dev]"
]

[project.urls]
Homepage = "https://github.com/pdfx-bench/pdfx-bench"
Documentation = "https://pdfx-bench.readthedocs.io"
Repository = "https://github.com/pdfx-bench/pdfx-bench"
Issues = "https://github.com/pdfx-bench/pdfx-bench/issues"

[project.scripts]
pdfx-bench = "pdfx_bench.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pdfx_bench*"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=pdfx_bench --cov-report=html --cov-report=term-missing"
